refactor(DemoTagOverText): display annotation result and remove toolbar
