+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/Demo/TagOverTextContainer.scss
@@ -8,6 +8,9 @@
     border-radius: 3px;
     //  margin-bottom: 50px;
   }
+  &__result{
+    overflow-wrap: break-word;
+  }
 }
 
 .konvajs-content__container--demo > div > .konvajs-content {
