--- src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/Demo/Toolbar.scss
@@ -1,4 +1,0 @@
-﻿
-.toolbar--demo{
-  position: relative;
-}
