+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/TagOverText/TagOverText.container.js
@@ -249,7 +249,6 @@ const TagOverTextContainer = ({ expectedOutput, url, onSubmit }) => {
     }*/
 
     if (image) {
-      console.log("expectedOutput: " + JSON.stringify(expectedOutput));
       const labels = expectedOutput.map((label, index) => {
         return {
           id: label.id,
